requirements.txt
discord.py
openai
python-dotenv
flask  # for keep_alive.py
